requirements.txt
flask
requests
numpy
Cython
Pillow
